Add a GrayScale effect and parallel processing support
With this change, there is now a Grayscale function that can convert
an input image to its grayscale representation.

Also the code is significantly faster as we use the Pix array of the
image.Image instances, vs. the At(x,y), Set(x,y) functions which are
a lot slower.

Finally, the code now runs across multiple goroutines, the image is
split into pieces to be processed by multiple goroutines in parallel,
parallel not concurrent, this brings significant speed improvements.