Add cartoon effect
The cartoon effect is achieved by performing edge detection on the
input image, then the input image is rendered as an oil paiting
and the edges are drawn on top of the oil painting in black.

You can modify all of the parameters, each picture is different and
one set of values won't work for all images, you will have to
experiement with the settings.

If you are noticing you are getting too much noise in the edge
detection phase, you can set teh blur parameters so that the input
image is blurred before running edge detection. This removes some of
the high frequency noise.