Add pixelate effect
The pixelates effect breaks the input image into blocks, based on the
block size passed in by the user, and then fills each block with the
average color of the pixels in the block, to give a pixelated effect.

One note is that the input image width and height must both divide
exactly by the specified block size.